pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "v0rtex"
version = "0.2.2-alpha.1"
description = "Dynamic JSON-based web scraper with anti-detection capabilities"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "v0rtex Team", email = "team@v0rtex.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "fake-useragent>=1.4.0",
    "undetected-chromedriver>=3.5.0",
    "cloudscraper>=1.2.71",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "asyncio-throttle>=1.0.2",
    "tenacity>=8.2.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.1",
    "jsonschema>=4.20.0",
    "cryptography>=41.0.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "easyocr>=1.7.0",
    "2captcha-python>=1.2.0",
    "anticaptchaofficial>=1.0.45",
    "requests-html>=0.10.0",
    "playwright>=1.40.0",
    "scrapy>=2.11.0",
    "scrapy-playwright>=0.0.34",
]

[project.entry-points."console_scripts"]
v0rtex = "v0rtex.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
vpn = [
    "requests[socks]>=2.31.0",
    "pysocks>=1.7.1",
    "stem>=1.8.0",
]
proxy = [
    "requests[socks]>=2.31.0",
    "pysocks>=1.7.1",
    "python-socks>=2.4.0",
]

[project.urls]
Homepage = "https://github.com/Apter-X/v0rtex"
Documentation = "https://v0rtex.readthedocs.io"
Repository = "https://github.com/Apter-X/v0rtex"
Issues = "https://github.com/Apter-X/v0rtex/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
